# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Crypto portfolio tracker and trading signal backend with AI-powered recommendations",
        "title": "Crypto Portfolio API",

        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/asset/{symbol}": {
            "get": {
                "description": "Get detailed information about a specific asset including current price, changes, and historical data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "Get asset detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset symbol (e.g., BTC, ETH)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AssetDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/asset/{symbol}/ai-signal": {
            "get": {
                "description": "Get DeepSeek AI-powered trading signal with natural language reasoning and recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "predictor"
                ],
                "summary": "Get AI-powered signal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset symbol (e.g., BTC, ETH)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Currency for price display (default: idr)",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AISignalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/asset/{symbol}/signal": {
            "get": {
                "description": "Get technical analysis signal based on RSI, SMA, and MACD indicators for a specific asset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "predictor"
                ],
                "summary": "Get trading signal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset symbol (e.g., BTC, ETH)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SignalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/news/briefing": {
            "get": {
                "description": "Get the latest crypto news briefing aggregated from RSS feeds with AI-powered sentiment analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get daily news briefing",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.DailyBriefing"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/portfolio": {
            "get": {
                "description": "Get aggregated portfolio with all assets, total net worth, and exchange rates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portfolio"
                ],
                "summary": "Get portfolio",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Portfolio"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.AIRecommendation": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "domain.AISignalResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "indicators": {
                    "$ref": "#/definitions/domain.SignalIndicators"
                },
                "reasoning": {
                    "type": "string"
                },
                "recommendation": {
                    "$ref": "#/definitions/domain.AIRecommendation"
                },
                "symbol": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/domain.AIUsage"
                }
            }
        },
        "domain.AIUsage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "cost_usd": {
                    "type": "number"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
        "domain.Asset": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_change_24h": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "value_idr": {
                    "type": "number"
                }
            }
        },
        "domain.AssetDetail": {
            "type": "object",
            "properties": {
                "changes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PricePoint"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price_idr": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "domain.DailyBriefing": {
            "type": "object",
            "properties": {
                "generated_at": {
                    "type": "string"
                },
                "news_count": {
                    "type": "integer"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sentiment": {
                    "type": "string"
                },
                "summary_date": {
                    "type": "string"
                }
            }
        },
        "domain.Portfolio": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Asset"
                    }
                },
                "exchange_rates": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "last_updated": {
                    "type": "string"
                },
                "total_net_worth_idr": {
                    "type": "number"
                }
            }
        },
        "domain.PricePoint": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "domain.SignalIndicators": {
            "type": "object",
            "properties": {
                "macd_histogram": {
                    "type": "number"
                },
                "macd_line": {
                    "type": "number"
                },
                "macd_signal": {
                    "type": "number"
                },
                "rsi": {
                    "type": "number"
                },
                "sma_200": {
                    "type": "number"
                },
                "sma_50": {
                    "type": "number"
                }
            }
        },
        "domain.SignalResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "generated_at": {
                    "type": "integer"
                },
                "indicators": {
                    "$ref": "#/definitions/domain.SignalIndicators"
                },
                "price": {
                    "type": "number"
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    }
}